Move content_utils endpoints to directory
