+++ src/endpoints/content_utils2/mod.rs
@@ -1,0 +1,6 @@
+pub mod list_content_templates;
+pub mod content_metadata_template;
+pub mod list_content_template_filenames;
+pub mod content_template;
+pub mod list_versifications;
+pub mod versification;
