+++ src/endpoints/mod.rs
@@ -3,10 +3,10 @@ pub mod navigation;
 pub mod clients;
 pub(crate) mod error;
 pub mod app_state;
-pub mod content_utils;
 pub mod git2;
 pub mod burrito2;
 pub mod gitea2;
 pub mod settings2;
 pub mod i18n2;
 pub mod atomics;
+pub mod content_utils2;
