+++ src/utils/launch.rs
@@ -84,12 +84,12 @@ pub(crate) fn add_routes(rocket_instance: Rocket<Build>) -> Rocket<Build> {
         .mount(
             "/content-utils",
             routes![
-                endpoints::content_utils::list_content_templates,
-                endpoints::content_utils::content_metadata_template,
-                endpoints::content_utils::list_content_template_filenames,
-                endpoints::content_utils::content_template,
-                endpoints::content_utils::list_versifications,
-                endpoints::content_utils::versification
+                endpoints::content_utils2::list_content_templates::list_content_templates,
+                endpoints::content_utils2::content_metadata_template::content_metadata_template,
+                endpoints::content_utils2::list_content_template_filenames::list_content_template_filenames,
+                endpoints::content_utils2::content_template::content_template,
+                endpoints::content_utils2::list_versifications::list_versifications,
+                endpoints::content_utils2::versification::versification
             ]
         )
         .mount(
